chore: complete remaining integrity review items (#211)
- Move test functions from 14 scorer files to src/eval_mm/tests/
- Resolve neologdn FIXME with try/except + NFKC fallback
- Document replace_images_tokens necessity in MMMU/JMMMU files
- Make Azure API version configurable via AZURE_API_VERSION env var

Fixes #194
