+++ src/eval_mm/metrics/jic_vqa_scorer.py
@@ -36,16 +36,3 @@ class JICVQAScorer(Scorer):
             result["average"] = 0.0
         output = AggregateOutput(result["average"], result)
         return output
-
-
-def test_jic_vqa_test():
-    refs = ["私は猫です。"]
-    preds = ["私は猫です。"]
-    from .scorer import ScorerConfig
-
-    scorer = JICVQAScorer(ScorerConfig(docs={"domain": ["test"]}))
-    scores = scorer.score(refs, preds)
-    assert scores == [1]
-    output = scorer.aggregate(scores)
-    assert output.overall_score == 1.0
-    assert output.details == {"test": 1.0, "average": 1.0}
