+++ src/eval_mm/metrics/llm_as_a_judge_scorer.py
@@ -84,65 +84,3 @@ class LlmAsaJudgeScorer(Scorer):
     def aggregate(scores: list) -> AggregateOutput:
         mean = sum(scores) / len(scores)
         return AggregateOutput(mean, {"llm_as_a_judge": mean})
-
-
-def test_llm_as_a_judge_scorer():
-    from eval_mm.utils.azure_client import MockChatAPI, OpenAIChatAPI
-
-    questions = ["What is the capital of Japan?", "What is the capital of France?"]
-    answers = ["Tokyo", "Paris"]
-    preds = ["", ""]
-    from .scorer import ScorerConfig
-
-    scorer = LlmAsaJudgeScorer(
-        ScorerConfig(docs={"input_text": questions}, client=MockChatAPI())
-    )
-    scores = scorer.score(answers, preds)
-    assert scores == [0, 0]
-    output = scorer.aggregate(scores)
-    assert output.overall_score == 0.0
-    assert output.details == {"llm_as_a_judge": 0.0}
-
-    import os
-
-    if os.getenv("AZURE_OPENAI_KEY"):
-        questions = ["What is the capital of Japan?", "What is the capital of France?"]
-        answers = ["Tokyo", "Paris"]
-        preds = ["Tokyo", "Paris"]
-        model_name = "gpt-4.1-2025-04-14"  # or any other model you want to test
-        scorer = LlmAsaJudgeScorer(
-            ScorerConfig(
-                docs={"input_text": questions},
-                client=OpenAIChatAPI(),
-                judge_model=model_name,
-            )
-        )
-        scores = scorer.score(answers, preds)
-        assert scores == [5, 5]
-
-        questions = [
-            "前年の合計所得が33万円以下の世帯の場合、軽減の割合は何割ですか。\n解答は数量のみで答えてください。",
-            "3つのラベルはどのような違いがありますか。\n解答は自由に記述してください。",
-            "国内の日本語教育の概要において、外国人に対する日本語教育の現状として、日本語教師数の数や学習者の数はどうなっていますか。\n解答は自由に記述してください。",
-        ]
-        answers = [
-            "7割です",
-            "下部に文字がプリントされているかどうか、プリントの内容が「CARBONNEUTRAL」または「CERTIFIEDMODEL」かの違いがあります。",
-            "国内の日本語教育の概要において、外国人に対する日本語教育の現状として、日本語教師数の数は34392人、学習者の数は139613人となっています。",
-        ]
-
-        preds = [
-            "7割",
-            "",
-            "日本語教師数の数は34392人、学習者の数は139613人となっています。",
-        ]
-        model_name = "gpt-4.1-2025-04-14"
-        scorer = LlmAsaJudgeScorer(
-            ScorerConfig(
-                docs={"input_text": questions},
-                client=OpenAIChatAPI(),
-                judge_model=model_name,
-            )
-        )
-        scores = scorer.score(answers, preds)
-        assert scores == [5, 1, 5]
